Added feature with the help of async/await
