+++ Assignments/assignment12/src/lib.rs
@@ -4,4 +4,4 @@ pub mod asynch {
     pub mod asynchronous;
 }
 
-pub mod synchronously;
+pub mod asynchronously;
